музыка добавлена(очень много откладок),работает
музыка добавлена(комментариев тьма)

+++ player.h
@@ -11,8 +11,8 @@ public:
     void invertGravity();
     void setSpeedMultiplier(qreal multiplier);
     void teleportTo(qreal x, qreal y);
-    qreal verticalSpeed = 0;
-    qreal getVerticalSpeed(){return verticalSpeed;}
+   float verticalSpeed = 0;
+ float getVerticalSpeed(){return verticalSpeed;}
     void resetVerticalSpeed(){verticalSpeed=0;};
     void setOnGround(bool on) { isOnGround = on; }
     bool isOnGroundf() const { return isOnGround; }
